创建menus，给project绑定管理员

+++ app/controllers/api/v1/projects_controller.rb
@@ -1,6 +1,8 @@
 class Api::V1::ProjectsController < ApplicationController
   def index
-    projects = Project.page(params[:page])
+    current_user = User.find request.env["current_user_id"]
+    return render status :unauthorized if current_user.nil?
+    projects = current_user.projects.page(params[:page])
               .per(params[:per])
 
     render json: {
@@ -14,7 +16,9 @@ class Api::V1::ProjectsController < ApplicationController
   end
 
   def show
-    project = Project.find_by_id params[:id]
+    current_user = User.find request.env["current_user_id"]
+    return render status :unauthorized if current_user.nil?
+    project = current_user.projects.find_by_id params[:id]
     return render status: :not_found if project.nil?
     render json: {
       resource: project
@@ -25,7 +29,7 @@ class Api::V1::ProjectsController < ApplicationController
     current_user = User.find request.env["current_user_id"]
     return render status :unauthorized if current_user.nil?
 
-    project = current_user.projects.create(name: params[:name])
+    project = current_user.projects.create name: params[:name], manager_user: current_user
 
     if project.save
         render json: { resource: project }, status: :ok
@@ -35,9 +39,13 @@ class Api::V1::ProjectsController < ApplicationController
   end
 
   def update
+    current_user = User.find request.env["current_user_id"]
+    return render status :unauthorized if current_user.nil?
     return render json: { success: false, message: "项目名称不能为空！" }, status: 422 if params[:name].nil?
     project = Project.find_by_id params[:id]
     return render status: :not_found if project.nil?
+
+    return render json: { message: "仅项目管理员允许修改！" },status: :forbidden if project.manager_user != current_user
 
     project.update params.permit(:name)
 
@@ -49,8 +57,12 @@ class Api::V1::ProjectsController < ApplicationController
   end
 
   def destroy
+    current_user = User.find request.env["current_user_id"]
+    return render status :unauthorized if current_user.nil?
     project = Project.find_by_id params[:id]
     return render status: :not_found if project.nil?
+
+    return render json: { message: "仅项目管理员允许删除！" },status: :forbidden if project.manager_user != current_user
 
     if project.destroy
       # 删除成功，返回成功信息
